requirements.txt
# Third-party dependencies (pip install -r requirements.txt)
pandas>=1.0.0          # Data processing
qrcode>=7.0.0          # QR code generation
opencv-python>=4.5.0   # Image processing (cv2 for image reading/processing)
pyzbar>=0.1.9          # QR code recognition (pyzbar.pyzbar for decoding)
Pillow>=9.0.0          # Image processing
matplotlib>=3.7.1      # Data visualization charts
pyotp>=2.9.0           # TOTP two-factor authentication
ttkthemes>=3.3.0       # Theme switching
web3>=6.0.0            # Ethereum blockchain interaction
eth-account>=0.8.0     # Ethereum account management
openpyxl>=3.0.0        # Excel file read/write support (required for pandas to_excel)
